feat(issue53): 調整 socket
